+++ changelog.md
@@ -2,8 +2,10 @@
 
 ## Upcoming
 * Increased Shovels knockback (0.8 -> 1)
-* Doubled the cooldown between fake mobs sound when tired
-  * And made it configurable
+* Tiredness
+  * Fake sounds times to play is now configurable
+  * Doubled the cooldown between fake mobs sound when tired
+    * And made it configurable
 
 ## 3.20.1
 * Fixed foggy weather overwriting Blindness